requirements.txt
Django==2.1.4
django-allauth==0.38.0
django-bootstrap3==11.0.0
django-ckeditor==5.6.1
django-debug-toolbar==1.11
django-js-asset==1.1.0
Pillow==5.3.0
pytz==2018.7
sqlparse==0.2.4
